Add more tests
Eliminate redundant CaptureDisposable call from expression trees
Split fakes into separate files
