+++ src/Microsoft.Extensions.DependencyInjection.Specification.Tests/Fakes/FakeDisposeCallback.cs
@@ -1,0 +1,12 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+
+namespace Microsoft.Extensions.DependencyInjection.Specification.Fakes
+{
+    public class FakeDisposeCallback
+    {
+        public List<object> Disposed { get; } = new List<object>();
+    }
+}
